+++ src/main/java/com/bjtu/afms/biz/impl/CommentBizImpl.java
@@ -1,9 +1,14 @@
 package com.bjtu.afms.biz.impl;
 
+import com.alibaba.fastjson.JSON;
 import com.bjtu.afms.biz.CommentBiz;
+import com.bjtu.afms.biz.LogBiz;
 import com.bjtu.afms.biz.PermissionBiz;
 import com.bjtu.afms.config.context.LoginContext;
 import com.bjtu.afms.enums.DataType;
+import com.bjtu.afms.enums.OperationType;
+import com.bjtu.afms.exception.BizException;
+import com.bjtu.afms.http.APIError;
 import com.bjtu.afms.http.Page;
 import com.bjtu.afms.model.Comment;
 import com.bjtu.afms.service.CommentService;
@@ -28,6 +33,9 @@ public class CommentBizImpl implements CommentBiz {
     @Resource
     private PermissionBiz permissionBiz;
 
+    @Resource
+    private LogBiz logBiz;
+
     @Override
     public Page<Comment> getCommentList(CommentQueryParam param, Integer page) {
         if (page == null) {
@@ -46,6 +54,8 @@ public class CommentBizImpl implements CommentBiz {
         comment.setUserId(LoginContext.getUserId());
         if (commentService.insertComment(comment) == 1) {
             permissionBiz.initResourceOwner(DataType.COMMENT.getId(), comment.getId(), comment.getUserId());
+            logBiz.saveLog(DataType.COMMENT, comment.getId(), OperationType.INSERT_COMMENT,
+                    null, JSON.toJSONString(comment));
             return true;
         } else {
             return false;
@@ -55,17 +65,37 @@ public class CommentBizImpl implements CommentBiz {
     @Override
     @Transactional
     public boolean modifyComment(int id, int score, String content) {
+        Comment old = commentService.selectComment(id);
+        if (old == null) {
+            throw new BizException(APIError.NOT_FOUND);
+        }
         Comment comment = new Comment();
         comment.setId(id);
         comment.setScore(score);
         comment.setContent(content);
-        return commentService.updateComment(comment) == 1;
+        if (commentService.updateComment(comment) == 1) {
+            logBiz.saveLog(DataType.COMMENT, comment.getId(), OperationType.UPDATE_COMMENT,
+                    JSON.toJSONString(old), JSON.toJSONString(comment));
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
     @Transactional
     public boolean deleteComment(int commentId) {
-        permissionBiz.deleteResource(DataType.COMMENT.getId(), commentId);
-        return commentService.deleteComment(commentId) == 1;
+        Comment old = commentService.selectComment(commentId);
+        if (old == null) {
+            throw new BizException(APIError.NOT_FOUND);
+        }
+        if (commentService.deleteComment(commentId) == 1) {
+            permissionBiz.deleteResource(DataType.COMMENT.getId(), commentId);
+            logBiz.saveLog(DataType.COMMENT, commentId, OperationType.DELETE_COMMENT,
+                    JSON.toJSONString(old), null);
+            return true;
+        } else {
+            return false;
+        }
     }
 }
